+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Audio and video layers going silent after the first time recording the movie ([#106](https://github.com/etro-js/etro/issues/106)).
 
 ### Security
-- Bump word-wrap from 1.2.3 to 1.2.5 (#222).
+- Bump word-wrap from 1.2.3 to 1.2.5 ([#222](https://github.com/etro-js/etro/issues/222)).
 
 ## [0.10.1] - 2023-07-16
 ### Security