+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.11.0] - 2023-08-05
 ### Added
 - `duration` option for `Movie#play` ([#208](https://github.com/etro-js/etro/pull/208)).
 
@@ -292,7 +292,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - Gaussian blur
   - Transform
 
-[Unreleased]: https://github.com/etro-js/etro/compare/v0.10.1...HEAD
+[0.11.0]: https://github.com/etro-js/etro/compare/v0.10.1...v0.11.0
 [0.10.1]: https://github.com/etro-js/etro/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/etro-js/etro/compare/v0.9.1...v0.10.0
 [0.9.1]: https://github.com/etro-js/etro/compare/v0.9.0...v0.9.1